CRDCDH-730 Refactor default validation selection
