+++ src/utils/index.ts
@@ -6,3 +6,4 @@ export * from './formModeUtils';
 export * from './profileUtils';
 export * from './historyUtils';
 export * from './dataModelUtils';
+export * from './dataValidationUtils';
